10-28 pm 4:32bc

+++ src/main/webapp/js/adminPages/boardList.js
@@ -4,7 +4,30 @@ var boardSearchValue = null;
 
 setTimeout(function() {
     var page;
-    loadBoardPage(page);
+
+    if (usercode1!=null &&usercode1!=0 &&usercode1!=""){
+        $.ajax({
+            url:"/adminPages/checkuser",
+            type:"post",
+            data:{
+                usercode:usercode1
+            },success:function(r){
+                var role=r.role;
+                if (role!="ROLE_USER"){
+                    loadBoardPage(page);
+                }else{
+                    window.location.href="/";
+                }
+
+
+            }
+        })
+
+    }else{
+        window.location.href="/";
+    }
+
+
 }, 200);
 
 function loadBoardPage(page, boardSearchType, boardSearchType2, boardSearchValue) {
@@ -16,7 +39,8 @@ function loadBoardPage(page, boardSearchType, boardSearchType2, boardSearchValue
     var BoardData = {
         page: page,
         searchKey: boardSearchType,
-        searchWord: boardSearchValue
+        searchWord: boardSearchValue,
+        usercode:usercode1
     };
     // 마라톤명 검색이 있을 경우
     if (boardSearchType === "marathon_name" && boardSearchValue) {
@@ -42,24 +66,40 @@ function loadBoardPage(page, boardSearchType, boardSearchType2, boardSearchValue
             console.log("Total Pages: ", r.pvo.totalPage);*/
             var BoardList = r.list;
             var pVO = r.pvo;
+            var roles=r.Avo.role;
+            var cerate=r.Avo.permission_add;
+            var deleted=r.Avo.permission_delete;
+            var edit=r.Avo.permission_edit;
+            console.log(roles);
             if (r.list.length > 0) {
                 // 결과값이 존재할 때 검색어 초기화
                 document.getElementById("searchtext").value = ""; // 검색어 입력 필드의 ID가 'searchInput'인 경우
             }
+            if (roles<3||r.pvo.admin_code==0){
+                if (roles<2||r.pvo.admin_code==0){
+                    var downloadbuttontag = `<input type="button" value="게시글 리스트 받기" onClick="excelDownload()"/>`;
+                    document.getElementById("downloadbutton").innerHTML = downloadbuttontag;
+                    document.getElementById("downloadbutton").style.display = "block";
+                }
+                //작성하기 버튼 생성
+                if (cerate=="1"||r.pvo.admin_code==0){
+                    var writeButtonHtml  = `<input type="button" value="글 작성하기" onClick="writePost()"/>`;
+                    document.getElementById("writebutton").innerHTML = writeButtonHtml; // 기존 버튼을 덮어씀
+                    document.getElementById("writebutton").style.display = "block";
+                }
 
-            var downloadbuttontag = `<input type="button" value="게시글 리스트 받기" onClick="excelDownload()"/>`;
-            document.getElementById("downloadbutton").innerHTML = downloadbuttontag;
-            document.getElementById("downloadbutton").style.display = "block";
-
-            //작성하기 버튼 생성
-            var writeButtonHtml  = `<input type="button" value="글 작성하기" onClick="writePost()"/>`;
-            document.getElementById("writebutton").innerHTML = writeButtonHtml; // 기존 버튼을 덮어씀
-            document.getElementById("writebutton").style.display = "block";
-
+            }
             var tag = "<li><div id='board_title2'><div class='board_code'>마라톤번호</div><div class='marathon_name'>마라톤명</div>";
             tag += "<div class='created_date'>생성날짜</div><div class='activation_date'>활성화 여부</div>";
             tag += "<div class='deleted_date'>삭제 여부</div><div class='deleted_date'>삭제 완료일</div>"; // 삭제 완료일 추가
-            tag += "<div class='edit_button'>수정</div><div class='delete_button'>삭제</div>";
+            if (roles<3||r.pvo.admin_code==0){
+                if (edit== "1"||r.pvo.admin_code==0){
+                    tag += "<div class='edit_button'>수정</div>";
+                }
+                if (deleted== "1" ||r.pvo.admin_code==0){
+                    tag +="<div class='delete_button'>삭제</div>";
+                }
+            }
             tag += "</div></li>";
             /*console.log("BoardData:", BoardData); // AJAX 호출 전에 추가*/
 
@@ -82,10 +122,15 @@ function loadBoardPage(page, boardSearchType, boardSearchType2, boardSearchValue
                 } else {
                     tag += "<div class='deleted_date'>N/A</div>";
                 }
-
-                // 수정 및 삭제 버튼
-                tag += "<div class='edit_button'><button class='btn btn-outline-success' onClick='editMarathon(" + board.marathon_code + ")'>수정</button></div>";
-                tag += "<div class='delete_button'><button class='btn btn-outline-danger' onClick='deleteBoard(" + board.marathon_code + ")'>삭제</button></div>";
+                if(roles<3){
+                    // 수정 및 삭제 버튼
+                    if (edit=="1"||r.pvo.admin_code==0){
+                        tag += "<div class='edit_button'><button class='btn btn-outline-success' onClick='editMarathon(" + board.marathon_code + ")'>수정</button></div>";
+                    }
+                    if (deleted=="1"||r.pvo.admin_code==0){
+                        tag += "<div class='delete_button'><button class='btn btn-outline-danger' onClick='deleteBoard(" + board.marathon_code + ")'>삭제</button></div>";
+                    }
+                }
                 tag += "</div></li>";
             });
